+++ CHANGELOG.md
@@ -6,6 +6,34 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 *Nothing here yet.*
 
+## [v0.4.0]
+- Officially support PyPy 5.6, including Python 2 and 3 support
+
+## [v0.3.12]
+- Improve support for nesting styles with `str.format()`
+
+## [v0.3.11]
+- Support original module functionality for module hack
+
+## [v0.3.10]
+- Add disable method
+
+## [v0.3.9]
+
+### Fixed
+- Augment add a `str()` to `ColorfulString`
+
+## [v0.3.8]
+
+### Added
+- Support creating unstyled ColorfulString
+
+## [v0.3.7]
+
+### Added
+- Support augmented add for ColorfulString's
+- Validate hex string for RGB colors
+
 ## [v0.3.6]
 
 ### Added
@@ -42,7 +70,14 @@ This release is just to fix the PyPI project page.
 
 - Initial release
 
-[Unreleased]: https://github.com/timofurrer/colorful/compare/v0.3.6...HEAD
+[Unreleased]: https://github.com/timofurrer/colorful/compare/v0.4.0...HEAD
+[v0.4.0]: https://github.com/timofurrer/colorful/compare/v0.3.12...v0.4.0
+[v0.3.12]: https://github.com/timofurrer/colorful/compare/v0.3.11...v0.3.12
+[v0.3.11]: https://github.com/timofurrer/colorful/compare/v0.3.10...v0.3.11
+[v0.3.10]: https://github.com/timofurrer/colorful/compare/v0.3.9...v0.3.10
+[v0.3.9]: https://github.com/timofurrer/colorful/compare/v0.3.8...v0.3.9
+[v0.3.8]: https://github.com/timofurrer/colorful/compare/v0.3.7...v0.3.8
+[v0.3.7]: https://github.com/timofurrer/colorful/compare/v0.3.6...v0.3.7
 [v0.3.6]: https://github.com/timofurrer/colorful/compare/v0.3.5...v0.3.6
 [v0.3.5]: https://github.com/timofurrer/colorful/compare/v0.3.4...v0.3.5
 [v0.3.4]: https://github.com/timofurrer/colorful/compare/v0.3.3...v0.3.4